--- results/benchmark.txt
@@ -1,187 +1,0 @@
-Method Time Sample
-
-# Benchmark SpacedOut-Set (erros repetidos)
-SpacedOut-Set 0.00059101 Palavra1
-SpacedOut-Set 0.00050076 Palavra1
-SpacedOut-Set 0.00049688 Palavra1
-SpacedOut-Set 0.00049901 Palavra1
-SpacedOut-Set 0.00051085 Palavra1
-SpacedOut-Set 0.00049968 Palavra1
-SpacedOut-Set 0.00050591 Palavra1
-SpacedOut-Set 0.00055768 Palavra1
-SpacedOut-Set 0.00057411 Palavra1
-SpacedOut-Set 0.00066678 Palavra1
-SpacedOut-Set 0.00075516 Palavra1
-SpacedOut-Set 0.00075664 Palavra1
-SpacedOut-Set 0.00036580 Palavra2
-SpacedOut-Set 0.00034920 Palavra2
-SpacedOut-Set 0.00034764 Palavra2
-SpacedOut-Set 0.00034583 Palavra2
-SpacedOut-Set 0.00038921 Palavra2
-SpacedOut-Set 0.00040805 Palavra2
-SpacedOut-Set 0.00039946 Palavra2
-SpacedOut-Set 0.00038509 Palavra2
-SpacedOut-Set 0.00043671 Palavra2
-SpacedOut-Set 0.00046260 Palavra2
-SpacedOut-Set 0.00055967 Palavra2
-SpacedOut-Set 0.00086276 Palavra3
-SpacedOut-Set 0.00074707 Palavra3
-SpacedOut-Set 0.00072958 Palavra3
-SpacedOut-Set 0.00076158 Palavra3
-SpacedOut-Set 0.00073966 Palavra3
-SpacedOut-Set 0.00075630 Palavra3
-SpacedOut-Set 0.00075116 Palavra3
-SpacedOut-Set 0.00080743 Palavra3
-SpacedOut-Set 0.00078968 Palavra3
-SpacedOut-Set 0.00078628 Palavra3
-SpacedOut-Set 0.00081656 Palavra3
-SpacedOut-Set 0.00085816 Palavra3
-SpacedOut-Set 0.00092962 Palavra3
-SpacedOut-Set 0.00086993 Palavra4
-SpacedOut-Set 0.00090704 Palavra4
-SpacedOut-Set 0.00092107 Palavra4
-SpacedOut-Set 0.00103401 Palavra4
-SpacedOut-Set 0.00101426 Palavra4
-SpacedOut-Set 0.00102612 Palavra4
-SpacedOut-Set 0.00102701 Palavra4
-SpacedOut-Set 0.00108903 Palavra4
-SpacedOut-Set 0.00108427 Palavra4
-SpacedOut-Set 0.00115604 Palavra4
-SpacedOut-Set 0.00121940 Palavra4
-SpacedOut-Set 0.00118601 Palavra4
-SpacedOut-Set 0.00124859 Palavra4
-SpacedOut-Set 0.00127628 Palavra4
-SpacedOut-Set 0.00131527 Palavra4
-SpacedOut-Set 0.00128051 Palavra4
-SpacedOut-Set 0.00087217 Palavra5
-SpacedOut-Set 0.00086545 Palavra5
-SpacedOut-Set 0.00085760 Palavra5
-SpacedOut-Set 0.00086703 Palavra5
-SpacedOut-Set 0.00086089 Palavra5
-SpacedOut-Set 0.00086985 Palavra5
-SpacedOut-Set 0.00087908 Palavra5
-SpacedOut-Set 0.00094951 Palavra5
-SpacedOut-Set 0.00093798 Palavra5
-SpacedOut-Set 0.00101915 Palavra5
-SpacedOut-Set 0.00105362 Palavra5
-SpacedOut-Set 0.00114245 Palavra5
-SpacedOut-Set 0.00086675 Palavra6
-SpacedOut-Set 0.00083753 Palavra6
-SpacedOut-Set 0.00085745 Palavra6
-SpacedOut-Set 0.00091928 Palavra6
-SpacedOut-Set 0.00092847 Palavra6
-SpacedOut-Set 0.00092950 Palavra6
-SpacedOut-Set 0.00092398 Palavra6
-SpacedOut-Set 0.00097609 Palavra6
-SpacedOut-Set 0.00096500 Palavra6
-SpacedOut-Set 0.00096837 Palavra6
-SpacedOut-Set 0.00094996 Palavra6
-SpacedOut-Set 0.00096642 Palavra6
-SpacedOut-Set 0.00095915 Palavra6
-SpacedOut-Set 0.00112180 Palavra6
-SpacedOut-Set 0.00124384 Palavra6
-SpacedOut-Set 0.00111052 Palavra6
-SpacedOut-Set 0.00085772 Palavra7
-SpacedOut-Set 0.00085482 Palavra7
-SpacedOut-Set 0.00084537 Palavra7
-SpacedOut-Set 0.00089000 Palavra7
-SpacedOut-Set 0.00089200 Palavra7
-SpacedOut-Set 0.00092513 Palavra7
-SpacedOut-Set 0.00088558 Palavra7
-SpacedOut-Set 0.00093751 Palavra7
-SpacedOut-Set 0.00095301 Palavra7
-SpacedOut-Set 0.00103750 Palavra7
-SpacedOut-Set 0.00108564 Palavra7
-SpacedOut-Set 0.00111246 Palavra7
-SpacedOut-Lista 0.00074141 Palavra1
-SpacedOut-Lista 0.00111852 Palavra1
-SpacedOut-Lista 0.00130646 Palavra1
-SpacedOut-Lista 0.00154084 Palavra1
-SpacedOut-Lista 0.00177913 Palavra1
-SpacedOut-Lista 0.00213201 Palavra1
-SpacedOut-Lista 0.00240500 Palavra1
-SpacedOut-Lista 0.00278452 Palavra1
-SpacedOut-Lista 0.00309604 Palavra1
-SpacedOut-Lista 0.00355401 Palavra1
-SpacedOut-Lista 0.00386300 Palavra1
-SpacedOut-Lista 0.00424208 Palavra1
-SpacedOut-Lista 0.00056604 Palavra2
-SpacedOut-Lista 0.00080504 Palavra2
-SpacedOut-Lista 0.00093928 Palavra2
-SpacedOut-Lista 0.00112151 Palavra2
-SpacedOut-Lista 0.00125079 Palavra2
-SpacedOut-Lista 0.00149639 Palavra2
-SpacedOut-Lista 0.00168195 Palavra2
-SpacedOut-Lista 0.00183189 Palavra2
-SpacedOut-Lista 0.00222988 Palavra2
-SpacedOut-Lista 0.00241818 Palavra2
-SpacedOut-Lista 0.00273303 Palavra2
-SpacedOut-Lista 0.00104704 Palavra3
-SpacedOut-Lista 0.00163529 Palavra3
-SpacedOut-Lista 0.00192997 Palavra3
-SpacedOut-Lista 0.00221629 Palavra3
-SpacedOut-Lista 0.00258633 Palavra3
-SpacedOut-Lista 0.00303540 Palavra3
-SpacedOut-Lista 0.00346382 Palavra3
-SpacedOut-Lista 0.00395972 Palavra3
-SpacedOut-Lista 0.00424569 Palavra3
-SpacedOut-Lista 0.00479251 Palavra3
-SpacedOut-Lista 0.00525875 Palavra3
-SpacedOut-Lista 0.00677400 Palavra3
-SpacedOut-Lista 0.00657936 Palavra3
-SpacedOut-Lista 0.00121225 Palavra4
-SpacedOut-Lista 0.00208915 Palavra4
-SpacedOut-Lista 0.00259381 Palavra4
-SpacedOut-Lista 0.00315824 Palavra4
-SpacedOut-Lista 0.00373281 Palavra4
-SpacedOut-Lista 0.00443922 Palavra4
-SpacedOut-Lista 0.00518325 Palavra4
-SpacedOut-Lista 0.00596505 Palavra4
-SpacedOut-Lista 0.00646271 Palavra4
-SpacedOut-Lista 0.00692357 Palavra4
-SpacedOut-Lista 0.00735184 Palavra4
-SpacedOut-Lista 0.00793940 Palavra4
-SpacedOut-Lista 0.00846387 Palavra4
-SpacedOut-Lista 0.00916570 Palavra4
-SpacedOut-Lista 0.00963827 Palavra4
-SpacedOut-Lista 0.01037596 Palavra4
-SpacedOut-Lista 0.00128569 Palavra5
-SpacedOut-Lista 0.00195449 Palavra5
-SpacedOut-Lista 0.00226956 Palavra5
-SpacedOut-Lista 0.00274409 Palavra5
-SpacedOut-Lista 0.00318184 Palavra5
-SpacedOut-Lista 0.00366377 Palavra5
-SpacedOut-Lista 0.00418772 Palavra5
-SpacedOut-Lista 0.00499218 Palavra5
-SpacedOut-Lista 0.00584401 Palavra5
-SpacedOut-Lista 0.00604318 Palavra5
-SpacedOut-Lista 0.00662763 Palavra5
-SpacedOut-Lista 0.00704762 Palavra5
-SpacedOut-Lista 0.00124798 Palavra6
-SpacedOut-Lista 0.00192835 Palavra6
-SpacedOut-Lista 0.00228367 Palavra6
-SpacedOut-Lista 0.00284413 Palavra6
-SpacedOut-Lista 0.00349846 Palavra6
-SpacedOut-Lista 0.00401861 Palavra6
-SpacedOut-Lista 0.00453526 Palavra6
-SpacedOut-Lista 0.00530717 Palavra6
-SpacedOut-Lista 0.00585294 Palavra6
-SpacedOut-Lista 0.00620948 Palavra6
-SpacedOut-Lista 0.00649839 Palavra6
-SpacedOut-Lista 0.00698367 Palavra6
-SpacedOut-Lista 0.00736459 Palavra6
-SpacedOut-Lista 0.00785487 Palavra6
-SpacedOut-Lista 0.00846656 Palavra6
-SpacedOut-Lista 0.00895574 Palavra6
-SpacedOut-Lista 0.00120525 Palavra7
-SpacedOut-Lista 0.00193417 Palavra7
-SpacedOut-Lista 0.00226202 Palavra7
-SpacedOut-Lista 0.00267276 Palavra7
-SpacedOut-Lista 0.00313573 Palavra7
-SpacedOut-Lista 0.00370986 Palavra7
-SpacedOut-Lista 0.00422849 Palavra7
-SpacedOut-Lista 0.00497429 Palavra7
-SpacedOut-Lista 0.00556311 Palavra7
-SpacedOut-Lista 0.00599863 Palavra7
-SpacedOut-Lista 0.00645830 Palavra7
-SpacedOut-Lista 0.00708859 Palavra7
